requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
firebase-admin==6.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
email-validator==2.1.0
Werkzeug==3.0.1

# Testing dependencies
pytest==8.0.0
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-asyncio==0.23.3
faker==22.0.0
requests-mock==1.11.0